add original assignment text
I thought it made sense to include the original assignment text so that it was clear what I was supposed to do in the first place. While copying these down I realized I should really... actually read. I also removed the LICENSE file because I am not sure if it's okay to use the MIT license with the original assignment text included.