Delete LICENSE
Firstly, I added the license on accident by misclicking. While I do love open-source and libre software, I am not sure how licensing would work given that the instructions, which I did not write and come from various sources including textbooks that are probably copyrighted, are included.

I may re-add a different license at some point, modified to account for this. Actually, I am a bit worried that it might be illegal for me to upload the instructions like this. :P

If that turns out to be the case, then OH NO!!! But, if possible, I will fix it by replacing those files with my own summaries of what they originally said or something like that.